All files with more than one class corrected
