+++ ApsimNG/EventArguments/AddArcEventArgs.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using Models.Management;
+
+namespace UserInterface.EventArguments.DirectedGraph
+{
+    public class AddArcEventArgs : EventArgs
+    {
+        public RuleAction Arc { get; set; }
+    }
+}
