+++ ApsimNG/EventArguments/BubbleChartArgs.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using APSIM.Interop.Visualisation;
 using Models.Management;
 
 namespace UserInterface.EventArguments.DirectedGraph
@@ -19,39 +18,6 @@ namespace UserInterface.EventArguments.DirectedGraph
         {
             Nodes = nodes;
             Arcs = arcs;
-        }
-    }
-    public class AddNodeEventArgs : EventArgs
-    {
-        public StateNode Node { get; set; }
-        public AddNodeEventArgs(StateNode node) => Node = node;
-    }
-    public class DelNodeEventArgs : EventArgs
-    {
-        public string nodeNameToDelete { get; set; }
-    }
-    public class AddArcEventArgs : EventArgs
-    {
-        public RuleAction Arc { get; set; }
-    }
-    public class DelArcEventArgs : EventArgs
-    {
-        public string arcNameToDelete { get; set; }
-    }
-    public class ObjectMovedArgs : EventArgs
-    {
-        /// <summary>
-        /// The object which has been moved.
-        /// </summary>
-        public DGObject MovedObject { get; set; }
-
-        /// <summary>
-        /// Constructor.
-        /// </summary>
-        /// <param name="obj">The object which has been moved.</param>
-        public ObjectMovedArgs(DGObject obj)
-        {
-            MovedObject = obj;
         }
     }
 }
