+++ ApsimNG/EventArguments/DelArcEventArgs.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace UserInterface.EventArguments.DirectedGraph
+{
+    public class DelArcEventArgs : EventArgs
+    {
+        public string arcNameToDelete { get; set; }
+    }
+}
