+++ ApsimNG/EventArguments/DelNodeEventArgs.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace UserInterface.EventArguments.DirectedGraph
+{
+    public class DelNodeEventArgs : EventArgs
+    {
+        public string nodeNameToDelete { get; set; }
+    }
+}
