+++ ApsimNG/Interfaces/TIntArgs.cs
@@ -1,0 +1,9 @@
+﻿namespace UserInterface.Interfaces
+{
+    using System;
+
+    public class TIntArgs : EventArgs
+    {
+        public int Value { get; set; }
+    }
+}
