+++ ApsimNG/Interfaces/TStringArgs.cs
@@ -1,0 +1,9 @@
+﻿namespace UserInterface.Interfaces
+{
+    using System;
+
+    public class TStringArgs : EventArgs
+    {
+        public string Name { get; set; }
+    }
+}
