+++ ApsimNG/Utility/FontInput.cs
@@ -1,0 +1,7 @@
+﻿namespace Utility
+{
+    internal class FontInput : InputAttribute
+    {
+        public FontInput(string name) : base(name) { }
+    }
+}
