+++ ApsimNG/Views/AllowCloseArgs.cs
@@ -1,0 +1,11 @@
+﻿namespace UserInterface.Views
+{
+    using System;
+
+    /// <summary>An event argument structure with a field for allow to close.</summary>
+    public class AllowCloseArgs : EventArgs
+    {
+        public bool AllowClose;
+    }
+
+}
