+++ ApsimNG/Views/CLEM/CLEMView.cs
@@ -142,14 +142,4 @@ namespace UserInterface.Views
             }
         }
     }
-
-    public class TabChangedEventArgs : EventArgs
-    {
-        public string TabName { get; set; }
-
-        public TabChangedEventArgs(string myString)
-        {
-            this.TabName = myString;
-        }
-    }
 }
