+++ ApsimNG/Views/CLEM/TabChangedEventArgs.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace UserInterface.Views
+{
+    public class TabChangedEventArgs : EventArgs
+    {
+        public string TabName { get; set; }
+
+        public TabChangedEventArgs(string myString)
+        {
+            this.TabName = myString;
+        }
+    }
+}
