+++ ApsimNG/Views/InputView.cs
@@ -107,12 +107,4 @@ namespace UserInterface.Views
             }
         }
     }
-
-    /// <summary>
-    /// A class for holding info about a begin drag event.
-    /// </summary>
-    public class OpenDialogArgs : EventArgs
-    {
-        public string FileName;
-    }
 }
