+++ ApsimNG/Views/OpenDialogArgs.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace UserInterface.Views
+{
+    /// <summary>
+    /// A class for holding info about a begin drag event.
+    /// </summary>
+    public class OpenDialogArgs : EventArgs
+    {
+        public string FileName;
+    }
+}
